Reorganize a little and fix linting issues.
