+++ README.md
@@ -1,2 +1,3 @@
 # cis-bidsify
-Tools for Singularity- and heudiconv-based BIDSification and anonymization of MRI data optimized for the FIU HPC.
+Tools for Singularity- and heudiconv-based BIDSification and anonymization of
+MRI data optimized for the FIU HPC.
